+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # Unreleased
 
 * Displayed warning when attempting to continue a workflow on unprocessed images (\#774);
+* Added "Get URL" button for plates on dataset page (\#774);
+* Fixed workflow status update issue (\#774);
 
 # 1.17.2